+++ CHANGELOG.md
@@ -7,3 +7,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Add manager
+- Add command class
+- Add init method
+- Add repository